requirements.txt
scikit-learn
spotipy
sentence-transformers
torch
matplotlib
minisom
librosa
numpy
pandas
yt-dlp
lyricsgenius
langdetect
--extra-index-url https://download.pytorch.org/whl/cpu
transformers
python-dotenv